+++ src/SpotifySearchQuery.svelte
@@ -41,6 +41,7 @@
 
 <script>
   import { restore, query } from "svelte-apollo";
+  import CheckLogin from './CheckLogin.svelte'
 
   export let cache, search;
   restore(client, SPOTIFY_SEARCH_QUERY, cache.data);
@@ -58,10 +59,6 @@
       request.refetch();
     }
   }
-
-  function needsLoginService(result) {
-    return auth.findMissingAuthServices(result.errors)[0];
-  }
 </script>
 
 <div>
@@ -70,17 +67,12 @@
     {#await $request}
       <li>Loading...</li>
     {:then result}
-      <pre>{JSON.stringify(result, null, 2)}</pre>
-      <button on:click={() => request.refetch({ fetchPolicy: 'network-only' })}>
-        Refetch
-      </button>
-
-      {#if needsLoginService(result)}
-        <button on:click={() => handleLogin(needsLoginService(result))}>
-          Login to
-          {needsLoginService(result)}
+      <CheckLogin {result}>
+        <pre>{JSON.stringify(result, null, 2)}</pre>
+        <button on:click={() => request.refetch({ fetchPolicy: 'network-only' })}>
+          Refetch
         </button>
-      {/if}
+      </CheckLogin>
     {:catch error}
       <li>Error loading articles: {error}</li>
       <button on:click={handleLogin}>Login</button>
